+++ src/components/organisms/AutoSuggestion/AutoSuggestionList.tsx
@@ -4,18 +4,12 @@ import { compute } from "compute-scroll-into-view";
 import { cn } from "@/common/utils/cn";
 import { CLASSPREFIX as eccgui } from "@/configuration/constants";
 
-import {
-    Highlighter,
-    Menu,
-    MenuItem,
-    OverflowText,
-    OverviewItem,
-    OverviewItemDescription,
-    OverviewItemLine,
-    Spacing,
-    Spinner,
-    Tooltip,
-} from "@/index";
+import { Spacing } from "@/components/atoms/Separation/Spacing";
+import { Spinner } from "@/components/atoms/Spinner/Spinner";
+import { Tooltip } from "@/components/atoms/Tooltip/Tooltip";
+import { Highlighter, OverflowText } from "@/components/atoms/Typography";
+import { Menu, MenuItem } from "@/components/molecules/Menu";
+import { OverviewItem, OverviewItemDescription, OverviewItemLine } from "@/components/molecules/OverviewItem";
 import { CodeAutocompleteFieldSuggestionWithReplacementInfo } from "./AutoSuggestion";
 
 export interface AutoSuggestionListProps extends Omit<React.HTMLAttributes<HTMLDivElement>, "children"> {