+++ src/components/organisms/Workspace/index.ts
@@ -1,3 +1,2 @@
 export * from "./WorkspaceContent";
 export * from "./WorkspaceMain";
-export * from "./WorkspaceSide";