+++ src/test/fileMock.js
@@ -1,0 +1,2 @@
+// jest stand-in for asset imports (webpack serves these via url-loader / asset modules)
+module.exports = "test-file-stub";